+++ src/Commands/Context/PrintContext.js
@@ -10,7 +10,7 @@ module.exports = function ( Engine )
 		CommandHelp: `$PrintContext:
 Prints a context variable value to the console.
 Fields:
-- context: The context variable to print.
+- context: (optional) The context variable to print. Leave empty to print entire context.
 `,
 
 
@@ -18,10 +18,14 @@ Fields:
 		Invoke: async function ( Step, Context )
 		{
 			if ( typeof Step === 'undefined' ) { throw new Error( `${Command.CommandName}: ${Command.CommandName}: The [Step] parameter is required.` ); }
-			if ( typeof Step.context === 'undefined' ) { throw new Error( `${Command.CommandName}: ${Command.CommandName}: The "context" field is required.` ); }
-			if ( typeof Step.context !== 'string' ) { throw new Error( `${Command.CommandName}: ${Command.CommandName}: The "context" field must be a string.` ); }
-			let value = Engine.Loose.GetObjectValue( Context, Step.context );
+
+			let value = Context;
+			if ( typeof Step.context === 'string' ) 
+			{
+				value = Engine.Loose.GetObjectValue( Context, Step.context );
+			}
 			console.log( JSON.stringify( value, null, '    ' ) );
+
 			return true;
 		},